Added condition for null pointer in deletefeedback.

+++ api/src/main/java/org/ednovo/gooru/domain/service/FeedbackServiceImpl.java
@@ -168,7 +168,8 @@ public class FeedbackServiceImpl extends BaseServiceImpl implements FeedbackServ
 		if(feedback == null){
 			throw new NotFoundException(generateErrorMessage(GL0056, FEEDBACK));
 		} else {
-			if (this.getUserManagementService().isContentAdmin(user) || feedback.getCreator().getPartyUid().equals(user.getPartyUid())) {
+			
+			if (this.getUserManagementService().isContentAdmin(user) || (feedback.getCreator()!=null && feedback.getCreator().getPartyUid().equals(user.getPartyUid()))) {
 				this.getFeedbackRepository().remove(feedback);
 			} else {
 				throw new UnauthorizedException(generateErrorMessage(GL0057, FEEDBACK));
